requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
dotenv
requests
azure-identity 
azure-mgmt-containerservice
kubernetes
redis 
flask-caching
psycopg2-binary
